fix(Dr. Rai Reports): Use one flag per site (#5709)
**Story card:**
[sc-16860](https://app.shortcut.com/simpledotorg/story/16860/enable-action-plans-are-visible-in-app-progress-tab)

## Because

When both tags are used, both tags need to be enabled for the user to
see the report in the progress tab. This is not necessary since the
action plans in the progress tab is view-only. Only depending on one tag
is sufficient in this case.

## This addresses

Using one feature flag only for the progress tab

## Test instructions

suite tests