Removing screened patients from reporting_patient_states calculations (#5721)
**Story card:**
[sc-17153](https://app.shortcut.com/simpledotorg/story/17153/update-reporting-patient-states-to-not-include-screened-patients)

## Because

We are adding screened patients to the system and don't want them to
affect our reporting numbers.

## This addresses

We are using "diagnosed_confirmed_at" column in the patients column to
segregate between screened and regular patients.
A null value in this would mean that patient is still under screening,
hence exclude them.
A timestamp value in this marks the time when the first diagnosis of
patient was confirmed - resulting him being considered as a "regular"
patient. This timestamp denotes the time from when we would actually
start calculating it's actual "registration" to the program.

## Test instructions

Test suite

---------

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>